pipeline/pyproject.toml
[project]
name = "pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Arterialist", email = "qazedcwsxrfv228@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.100.0",
    "h5py>=3.7.0",
    "httpx>=0.24.0",
    "jinja2>=3.1.0",
    "loguru>=0.7.0",
    "matplotlib>=3.7.0",
    "networkx>=3.0.0",
    "numpy>=2.0.0",
    "pandas>=2.0.0",
    "plotly>=5.13.0",
    "pydantic>=2.0.0",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "python-multipart>=0.0.6",
    "pyyaml>=6.0.0",
    "requests>=2.28.0",
    "scikit-learn>=1.2.0",
    "scipy>=1.10.0",
    "seaborn>=0.12.0",
    "snntorch>=0.9.1",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "tqdm>=4.64.0",
    "uvicorn>=0.23.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = false
